streamlit_app/requirements.txt
pandas==1.1.0
numpy==1.19.1
Pillow==7.2.0
scikit_learn==0.23.2
streamlit==0.68.0
